backend/requirements.txt
flask==1.1.1
flask-mongoengine
flask-restx
flask_cors
flask-bcrypt
Flask-JWT-Extended
flask-mail
gunicorn==20.0.4
Werkzeug==0.16.1

